Add Memfault to the "Crash and Exception Handling" section
Thank you for contributing to this list. This list is for **SaaS**
services that offer a **free tier** to help developers evaluate and
build something that users can later use and get support for.

The focus of this list is quite broad but we try to keep things
limited to that which infrastructure developers, like DevOps Practitioners,
would find useful.

We do not anymore accept cPanel like PHP+MySQL hosting services.

<!-- Only for new submissions -->

Please ensure your submission ticks all of these boxes:

 - [x] This is Software as a Service not self hosted
 - [x] It has a free tier not just a free trial
 - [x] The submission mentions what is free
 - [x] The submission is not already present in the list

This list is the result of more than a thousand people contributing
to make something useful, we appreciate your efforts.